Added ApiKey security to all endpoints
Added ApiKey security to all endpoints. Now when the app starts and the user wants to do CRUD operations via selected endpoints, he will be shown "API Key missing". The user first needs to click on the "Authorize" button in the upper right corner, enter the ApiKey and then he will be "logged in" and will be able to do CRUD operations as before.